Check for existing zulu installation when installing Script will exit if the zulu base directory is found.
If the zulu config directory is found, the user is prompted whether or not to
overwrite the existing config. If they choose no, zulu will continue the install
leaving the existing config intact. This is handy if the config dir is synced
between computers.

Also runs zulu bundle at the end of the install, so that if a packagefile is
found all packages within it are installed and available immediately as part
of the install.